% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_code.R, R/download_solution.R
\name{download_code}
\alias{download_code}
\title{A function that downloads simulator code}
\usage{
download_code(modelsettings, modelfunction)

download_code(modelsettings, modelfunction)
}
\arguments{
\item{modelsettings}{A list with model settings.
#' Needs to contain list elements with names and values for all inputs expected 
by simulation function. Required are:
Name of simulation function in variable modelsettings$simfunction
Also needs to contain an element plotscale 
to indicate which axis should be on a log scale (x, y or both), 
a list element nplots to indicate number of plots that should be produced
when calling the generate_plot function with the result, 
and a list element modeltype which specifies what kind of model should be run. 
Currently one of (_ode_, _discrete_, _stochastic_, _usanalysis_, _modelexploration_, _fit_ ). Stochastic models also need an nreps list entry to indicate numer of repeat simulations.}

\item{modelfunction}{The name of the simulator function being called. The name
must be one of the simulator functions in the DSAIDE package.}

\item{modelsettings}{A list with model settings.
#' Needs to contain list elements with names and values for all inputs expected 
by simulation function. Required are:
Name of simulation function in variable modelsettings$simfunction
Also needs to contain an element plotscale 
to indicate which axis should be on a log scale (x, y or both), 
a list element nplots to indicate number of plots that should be produced
when calling the generate_plot function with the result, 
and a list element modeltype which specifies what kind of model should be run. 
Currently one of (_ode_, _discrete_, _stochastic_, _usanalysis_, _modelexploration_, _fit_ ). Stochastic models also need an nreps list entry to indicate numer of repeat simulations.}

\item{modelfunction}{The name of the simulator function being called. The name
must be one of the simulator functions in the DSAIDE package.}
}
\value{
Creates an R script that runs the simulation specified in the app
and returns the text and plots created by the simulation.

Creates an R script that runs the simulation specified in the app
and returns the text and plots created by the simulation.
}
\description{
This function is called when the
user clicks the "Download Code" button in
the simulator app. It creates an R file which
will run the simulation specified in the
app from which the button is clicked, and
after running the simulation, returns the
plot and text created by generate_plots()
and generate_text().

This function is called when the
user clicks the "Download Code" button in
the simulator app. It creates an R file which
will run the simulation specified in the
app from which the button is clicked, and
after running the simulation, returns the
plot and text created by generate_plots()
and generate_text().
}
