% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAIDE-package.R
\docType{package}
\name{DSAIDE-package}
\alias{DSAIDE}
\alias{DSAIDE-package}
\title{DSAIDE: A package to learn about Dynamical Systems Approaches to Infectious
Disease Epidemiology}
\description{
This package provides a number of shiny apps that simulate various
infectious disease dynamics models By manipulating the models and working
through the instructions provided within the shiny UI, you can learn about
some important concepts in infectious disease epidemiology. You will also
learn how models can be used to study such concepts.
}
\details{
Start the main menu of the package by calling dsaidemenu().

To learn more about how to use the package, see the documentation
on the package website:
https://andreashandel.github.io/DSAIDE/
}
\seealso{
Useful links:
\itemize{
  \item \url{https://andreashandel.github.io/DSAIDE/}
  \item \url{https://github.com/andreashandel/DSAIDE/}
  \item Report bugs at \url{https://github.com/andreashandel/DSAIDE/issues}
}

}
\author{
\strong{Maintainer}: Andreas Handel \email{andreas.handel@gmail.com} (\href{https://orcid.org/0000-0002-4622-1146}{ORCID})

Other contributors:
\itemize{
  \item Cody Dailey [contributor]
  \item Isaac Chun-Hai Fung [contributor]
  \item Ronald Galiwango [contributor]
  \item Salil Goyal [contributor]
  \item Spencer Hall [contributor]
  \item Ben Listyg [contributor]
  \item Brian McKay [contributor]
  \item John Rossow [contributor]
  \item Sina Solaimanpour [contributor]
  \item Alexis Vittengl [contributor]
  \item Henok Woldu [contributor]
}

}
\keyword{internal}
