% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsaidemenu.R
\name{dsaidemenu}
\alias{dsaidemenu}
\title{The main menu for the DSAIDE package}
\usage{
dsaidemenu()
}
\description{
This function opens a menu that will allow the user to run the different simulation apps
}
\details{
Run this function with no arguments to start the main menu for the DSAIDE shiny apps
}
\examples{
\dontrun{dsaidemenu()}
}
\author{
Andreas Handel, Sina Solaimanpour
}

