% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flu1918data}
\alias{flu1918data}
\title{1918 Influenza mortality data}
\format{
A data frame/tibble with these variables:
\describe{
\item{Date}{Week of reporting, date format}
\item{Deaths}{New deaths in NYC per week per 100,000, numeric}
}
}
\usage{
flu1918data
}
\description{
Weekly influenza deaths per 100,000 during the 1918 pandemic for New York City.
}
\details{
Data is from Supplementary Table 1 of Mills et al 2004 Nature:
https://www.nature.com/articles/nature03063 

See this article and citations therein for more details on the data.
Note that only a subset of the data is present here and the data 
are meant to be used for illustrative purposes only. 
For a proper re-analysis of these data, use the source mentioned above or 
check out data available through project Tycho: https://www.tycho.pitt.edu/
}
\keyword{datasets}
