% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_shinyinput.R
\name{generate_shinyinput}
\alias{generate_shinyinput}
\title{A helper function that takes a model and generates shiny UI elements}
\usage{
generate_shinyinput(
  model_function,
  is_mbmodel,
  otherinputs = NULL,
  packagename
)
}
\arguments{
\item{model_function}{a string containing the name of a simulator function for which to build inputs}

\item{is_mbmodel}{a logical, indicating if the simulator function follows the modelbuilder syntax}

\item{otherinputs}{a text string that specifies a list of other shiny inputs to include in the UI}

\item{packagename}{name of package using this function (DSAIDE or DSAIRM)}
}
\value{
A renderUI object that can be added to the shiny output object for display in a Shiny UI
}
\description{
This function generates shiny UI inputs for a supplied model.
This is a helper function called by the shiny app.
}
\details{
This function is called by the Shiny app to produce the Shiny input UI elements.
model_function is assumed to be the name of a function.
The formals of the function will be parsed to create UI elements.
Therefore, all simulator_ R functions/scripts need to follow a specific syntax.
Either it needs to be created by modelbuilder and have vars/pars/times as vector inputs.
Or it has all single var = X inputs. They all need to have defaults provided.
Non-numeric arguments of functions are removed and need to be included in the otherinputs argument.
}
