% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{norodata}
\alias{norodata}
\title{Cases of norovirus during an outbreak}
\format{
A data frame with these variables:
\describe{
\item{Date}{Day of outbreak, all in December 2007, numeric}
\item{Cases}{New cases for the specified date, numeric}
}
}
\usage{
norodata
}
\description{
Norovirus case data from an outbreak among children on a school trip
}
\details{
The data are from Kuo 2009 Wien Klin Woch: 
"A non-foodborne norovirus outbreak among school children during a skiing holiday, Austria, 2007"
Specifically, the data comes from figure 1 of this article.
The total number of susceptibles was 284.

See this article and citations therein for more details on the data.
Note that only a subset of the data is present here and the data 
are meant for illustrative purposes only. 
For a proper re-analysis of these data, use the source mentioned above.
}
\keyword{datasets}
