% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_multisimoutput.R
\name{produce_multisimoutput}
\alias{produce_multisimoutput}
\title{A helper function that takes result from the simulators and produces plots and text output}
\usage{
produce_multisimoutput(input, output, res)
}
\value{
output a list with plot, text and warn elements for display in a shiny UI
}
\description{
This function generates plots and text to be displayed in the Shiny UI. 
This is a helper function. This version can process multiple simulation runs, supplied as a list
}
\details{
This function is called by the shiny server to produce output returned to the UI
}
\author{
Andreas Handel
}
\keyword{internal}

