% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_covid_stochastic.R
\name{simulate_covid_stochastic}
\alias{simulate_covid_stochastic}
\title{Stochastic simulation of a compartmental SEIR-type model for COVID-19}
\usage{
simulate_covid_stochastic(
  S = 1000,
  E = 10,
  beta_d = 5e-04,
  beta_u = 0.00025,
  beta_e = 1e-04,
  beta_red_factor = 0.5,
  t_int1 = 12,
  sigma = 0.2,
  gamma_u = 0.1,
  gamma_d = 0.5,
  t_int2 = 12,
  detect_frac_0 = 0.1,
  detect_frac_1 = 0.2,
  t_int3 = 12,
  tmax = 100,
  rngseed = 100
)
}
\arguments{
\item{S}{: initial number of susceptible hosts : numeric}

\item{E}{: initial number of infected, pre-symptomatic hosts : numeric}

\item{beta_d}{: level/rate of infectiousness for hosts in the I compartment : numeric}

\item{beta_u}{: level/rate of infectiousness for hosts in the Iu compartment : numeric}

\item{beta_e}{: level/rate of infectiousness for hosts in the Ie compartment : numeric}

\item{beta_red_factor}{: factor by which intervention reduces transmission : numeric}

\item{t_int1}{: time at which intervention reduces transmission : numeric}

\item{sigma}{: rate at which host leaves the E stage, inverse is pre-symptomatic duration : numeric}

\item{gamma_u}{: rate at which host goes through I stage, inverse is time to diagnosis, pre-intervention : numeric}

\item{gamma_d}{: rate at which host goes through I stage, inverse is time to diagnosis, post-intervention : numeric}

\item{t_int2}{: time at which intervention reduces time to diagnosis : numeric}

\item{detect_frac_0}{: fraction of infected individuals who will get diagnosed, pre-intervention : numeric}

\item{detect_frac_1}{: fraction of infected individuals who will get diagnosed, post-intervention : numeric}

\item{t_int3}{: time at which intervention reduces time to diagnosis : numeric}

\item{tmax}{: maximum simulation time : numeric}

\item{rngseed}{: seed for random number generator to allow reproducibility : numeric}
}
\value{
This function returns the simulation result as obtained from a call
  to the pomp::simulate integrator in list form. The list element ts is a
  dataframe where the first column is "time," and the remaining columns are the variables
}
\description{
Simulation of a stochastic SEIR model with the following compartments:
Susceptibles (S), latently infected (E), infected and later diagnosed (I), infected and never diagnosed (Iu),
Diagnosed cases (C) and non-diagnosed recovered (Ru)
}
\details{
A compartmental ID model with several states/compartments
is simulated as a stochastic model using the pomp simulat function
See the manual of this package for more details.
multiple dummy categories for E, I and Iu are implemented to get more realistic distributions.
Those dummy compartments are not publicly accessible.
Different interventions are implemented at different times.
}
\section{Warning}{

This function does not perform any error checking. So if you try to do
something nonsensical (e.g. have negative values or fractions > 1),
the code will likely abort with an error message
}

\examples{
# To run the simulation with default parameters just call the function:
result <- simulate_covid_stochastic()
# To choose parameter values other than the standard one, specify them like such:
result <- simulate_covid_stochastic(S = 2000,  sigma = 0.1, tmax = 200)
# You should then use the simulation result returned from the function, like this:
plot(result$ts[ , "time"],result$ts[ , "S"],xlab='Time',ylab='Number Susceptible',type='l')
}
\references{
See the manual for the pomp package for details on the algorithm.
}
\author{
Andreas Handel
}
