% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_transmissionmodes.R
\name{simulate_directtransmission}
\alias{simulate_directtransmission}
\title{simulate_directtransmission function}
\usage{
simulate_directtransmission(PopSize = 1e+06, I0 = 1, R0 = 0, tmax = 5,
  gamma = 13, beta.d = 4e-05, beta.f = 40, mu = 0 * 1/50, w = 0,
  k = 1e-06, scenario = 2)
}
\arguments{
\item{PopSize}{try a range of values for this and see how it affects the outbreak - this corresponds to the "initial-population" slider in NetLogo}

\item{I0}{initial number of infected hosts, again try a number of different values - this corresponds to the "initial-infected" slider in NetLogo}

\item{R0}{initial number of recovered hosts - for now we assume there are no recovered (immune) hosts. We get to this in the course.}

\item{tmax}{maximum simulation time, units of years, (equivalent to Max-time slider in the NetLogo simulation)}

\item{gamma}{rate at which a person leaves the infectious compartment, which is the inverse of the average duration of the infections period. Note: 1/gamma is the duration of the infectious period in units of YEARS so e.g. gamma=26 means infectious period = 1/26 years = 14 days}

\item{beta.d}{rate of transmission of pathogen from infected to susceptible host (density)}

\item{beta.f}{rate of transmission of pathogen from infected to susceptible host (frequency)}

\item{mu}{death rate parameter - inverse, i.e. 1/mu, is the average lifespan (in years), a value of 0 means no deaths. A value of 1/20 means e.g. 20 years lifespan}

\item{w}{duration of immunity. Setting it to zero means infinite immunity (no waning immunity), 1/w is duration of immunity, so e.g. w=0.5 means 2 years of immunity}

\item{k}{constant of proportionality for population-size dependent area - choose such that results look decent}

\item{scenario}{set to different values for different stransmission scenarios: 1 = density-dependent, 2 = frequency-dependent}
}
\value{
This function runs the and ODE solver and will return the frame that contains the data from the ODE solver
}
\description{
ESCRIPTION DESCRIPTION DESCRIPTION DESCRIPTION DESCRIPTION DESCRIPTION
}
\details{
Add details here...
}
\examples{
# To run the simulation with default parameters just call this function
simulate_directtransmission()
# One can also pass any number of the parameters as he needs
simulate_directtransmission(PopSize = 1e9)
simulate_directtransmission(PopSize = 1e9, gamma = 16)
}
\author{
Andreas Handel
}
\references{
We can add (a) reference(s) here..
}

