% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_introduction.R
\name{simulate_introduction}
\alias{simulate_introduction}
\title{Simulator for a basic SIR model}
\usage{
simulate_introduction(PopSize = 1000, I0 = 1, tmax = 300, gamma = 0.5,
  beta = 1/1000)
}
\arguments{
\item{PopSize}{specifies the initial number of individuals (suceptibles +
infected)}

\item{I0}{initial number of infected hosts}

\item{tmax}{maximum simulation time, units depend on choice of units for your
parameters}

\item{gamma}{rate at which a person leaves the infectious compartment, which
is the inverse of the average duration of the infections period}

\item{beta}{level of infectiousness, i.e. rate of transmission of pathogen
from infected to susceptible host}
}
\value{
The function returns the output from the odesolver as a matrix,
with one column per compartment/variable. The first column is time.
}
\description{
This function runs a simulation of a basic SIR model
using a set of 3 ordinary differential equations.
The user provides initial conditions and parameter values for the system.
The function simulates the ODE using an ODE solver from the deSolve package.
The function returns a matrix containing time-series of each variable and time.
}
\details{
A simple SIR model is simulated as a set of ordinary differential
equations, using an ode solver from the deSolve package.#'
}
\section{Warning}{

This function does not perform any error checking. So if you try to do
something nonsensical (e.g. have I0 > PopSize or any negative values), the
code will likely abort with an error message
}
\examples{
# To run the simulation with default parameters just call this function
result <- simulate_introduction()
# To choose parameter values other than the standard one, specify them e.g. like such
result <- simulate_introduction(PopSize = 2000, I0 = 10, tmax = 100, gamma = 1, beta = 1/100)
# You should then use the simulation result returned from the function, e.g. like this:
plot(result[,1],result[,2],xlab='Time',ylab='Number Susceptible',type='l')
}
\author{
Andreas Handel
}
\references{
See e.g. Keeling and Rohani 2008 for SIR models and the
  documentation for the deSolve package for details on ODE solvers
}

