% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_multipathogen.R
\name{simulate_multipathogen}
\alias{simulate_multipathogen}
\title{Simulation of a compartmental infectious disease transmission model with 2 types of hosts}
\usage{
simulate_multipathogen(S = 1000, I10 = 1, I20 = 0, I12 = 0, R1 = 0,
  R2 = 0, R12 = 0, tmax = 120, a = 0, b1 = 0, b2 = 0, b12 = 0,
  g1 = 1, g2 = 1, g12 = 1)
}
\arguments{
\item{S}{initial total number of hosts}

\item{tmax}{maximum simulation time, units of months}

\item{b1}{rate of transmission of type 1 pathogen to host}

\item{b2}{rate of transmission of type 2 pathogen to host}

\item{b12}{rate of transmission from infected type 2 host to susceptible type 1 host}

\item{g1}{the rate at which infected type 1 hosts recover}

\item{g2}{the rate at which infected type 2 hosts recover}

\item{g12}{the rate at which hosts infected with type 1 and 2 recover}

\item{I1}{initial number of hosts infected with type 1}

\item{I2}{initial number of hosts infected with type 2}
}
\value{
This function returns the simulation result as obtained from a call
  to the deSolve ode solver
}
\description{
This model allows for the simulation of an ID with 2 types of hosts
}
\details{
A compartmental ID model with several states/compartments
  is simulated as a set of ordinary differential
  equations. The function returns the output from the odesolver as a matrix,
  with one column per compartment/variable. The first column is time.
}
\section{Warning}{

  This function does not perform any error checking. So if you try to do
  something nonsensical (e.g. any negative values or fractions > 1),
  the code will likely abort with an error message
}

\examples{
  # To run the simulation with default parameters just call this function
  result <- simulate_multipathogen()
  # To choose parameter values other than the standard one, specify them e.g. like such
  result <- simulate_multipathogen(S10 = 100, S20 = 1e3,  tmax = 100)
  # You should then use the simulation result returned from the function, e.g. like this:
  plot(result[,1],result[,2],xlab='Time',ylab='Number Susceptible',type='l')
}
\references{
See e.g. Keeling and Rohani 2008 for SIR models and the
  documentation for the deSolve package for details on ODE solvers
}
\seealso{
The UI of the shiny app 'Multi-Pathogen Dynamics', which is part of this package, contains more details on the model
}
\author{
Andreas Handel and Spencer Hall
}
