% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_schoolclosure_ode.R
\name{simulate_schoolclosure_ode}
\alias{simulate_schoolclosure_ode}
\title{Simulation of a compartmental infectious disease transmission model with 3 types of hosts and intervention}
\usage{
simulate_schoolclosure_ode(
  Sc = 1000,
  Ic = 1,
  Sa = 1000,
  Ia = 1,
  Ss = 1000,
  Is = 0,
  bcc = 0.001,
  bca = 0.001,
  bcs = 0.001,
  baa = 0.001,
  bac = 0.001,
  bas = 0.001,
  bsc = 0.001,
  bsa = 0.001,
  bss = 0.001,
  gc = 0.1,
  ga = 0.1,
  gs = 0.1,
  dc = 0.001,
  da = 0.01,
  ds = 0.1,
  c_start = 30,
  c_end = 60,
  c_strength = 0.5,
  c_group = 1,
  tnew = 30,
  tmax = 365
)
}
\arguments{
\item{Sc}{: initial number of susceptible children : numeric}

\item{Ic}{:  initial number of infected children : numeric}

\item{Sa}{:  initial number of susceptible adults : numeric}

\item{Ia}{:  initial number of infected adults : numeric}

\item{Ss}{:  initial number of susceptible seniors : numeric}

\item{Is}{:  initial number of infected seniors : numeric}

\item{bcc}{:  rate of transmission from infected child to susceptible child : numeric}

\item{bca}{:  rate of transmission from infected child to susceptible adult : numeric}

\item{bcs}{:  rate of transmission from infected child to susceptible senior : numeric}

\item{baa}{:  rate of transmission from infected adult to susceptible adult : numeric}

\item{bac}{:  rate of transmission from infected adult to susceptible child : numeric}

\item{bas}{:  rate of transmission from infected adult to susceptible senior : numeric}

\item{bsc}{:  rate of transmission from infected senior to susceptible child : numeric}

\item{bsa}{:  rate of transmission from infected senior to susceptible adult : numeric}

\item{bss}{:  rate of transmission from infected senior to susceptible senior : numeric}

\item{gc}{:  rate at which infected children recover or die : numeric}

\item{ga}{:  rate at which infected adults recover or die : numeric}

\item{gs}{:  rate at which infected seniors recover or die : numeric}

\item{dc}{:  fraction of infected children who die : numeric}

\item{da}{:  fraction of infected adults who die : numeric}

\item{ds}{:  fraction of infected seniors who die : numeric}

\item{c_start}{: start of intervention : numeric}

\item{c_end}{: end of intervention : numeric}

\item{c_strength}{: strength of intervention, between 0 and 1 : numeric}

\item{c_group}{: target group for intervention (1 = children, 2 = adults, 3 = seniors) : numeric}

\item{tnew}{: time at which new infected adults enter : numeric}

\item{tmax}{:  maximum simulation time, units of months : numeric}
}
\value{
This function returns the simulation result as obtained from a call
  to the deSolve ode solver.
}
\description{
This model allows for the simulation of an ID with 3 types of hosts.
Groups are assumed to be children, adults and seniors.
Intervention can be applied to any of the groups for a certain duration.
}
\details{
A compartmental ID model with several states/compartments
  is simulated as a set of ordinary differential
  equations. The function returns the output from the odesolver as a matrix,
  with one column per compartment/variable. The first column is time.
  The model implement basic processes of infection, recovery and death.
  Treatment is applied, which reduces b by the indicated proportion, during times tstart and tend.
  At time intervals given by tnew, a new infected individual enters the population.
  The simulation also monitors the number of infected and when they drop below 1, they are set to 0.
}
\section{Warning}{

  This function does not perform any error checking. So if you try to do
  something nonsensical (e.g. any negative values or fractions > 1),
  the code will likely abort with an error message.
}

\examples{
  # To run the simulation with default parameters just call the function:
  result <- simulate_schoolclosure_ode()
}
\author{
Andreas Handel
}
