% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_SEIRS_model_stochastic.R
\name{simulate_SEIRS_model_stochastic}
\alias{simulate_SEIRS_model_stochastic}
\title{SEIRS model}
\usage{
simulate_SEIRS_model_stochastic(
  vars = c(S = 1000, E = 1, I = 1, R = 0),
  pars = c(bE = 0, bI = 0.001, gE = 1, gI = 1, w = 1, n = 0, m = 0),
  times = c(tstart = 0, tfinal = 100, dt = 0.1),
  rngseed = 123
)
}
\arguments{
\item{vars}{vector of starting conditions for model variables : numeric 
\itemize{ 
\item S : starting value for Susceptible : numeric
\item E : starting value for Exposed : numeric
\item I : starting value for Infected and Symptomatic : numeric
\item R : starting value for Recovered : numeric
}}

\item{pars}{vector of values for model parameters : numeric 
\itemize{ 
\item bE : level of infectiousness of exposed : numeric
\item bI : level of infectiousnes of symptomatic : numeric
\item gE : rate at which hosts move from E to I (inverse of exposed duration) : numeric
\item gI : rate at which hosts recover (inverse of symptomatic duration) : numeric
\item w : rate at which immunity wanes (inverse of immunity duration) : numeric
\item n : births : numeric
\item m : natural deaths (inverse of life-span) : numeric
}}

\item{times}{vector of values for model times : numeric 
\itemize{ 
\item tstart : Start time of simulation : numeric
\item tfinal : Final time of simulation : numeric
\item dt : Time step : numeric
}}
}
\value{
The function returns the output as a list. 
The time-series from the simulation is returned as a dataframe saved as list element \code{ts}. 
The \code{ts} dataframe has one column per compartment/variable. The first column is time.
}
\description{
A 4 compartment model with natural births and deaths and waning immunity.
}
\details{
The model includes susceptible, exposed/asymptomatic, infected/symptomatic, and recovered compartments. The processes that are modeled are infection, progression to infectiousness, recovery and waning immunity. Natural births and deaths are also included.

This code is was generated by the modelbuilder package R.  
The model is implemented as a set of stochastic equations using the adaptivetau package. 
The following R packages need to be loaded for the function to work: adpativetau
}
\section{Warning}{
 This function does not perform any error checking. So if you try to do something nonsensical (e.g. have negative values for parameters), the code will likely abort with an error message.
}

\section{Model Author}{
 Andreas Handel
}

\section{Model creation date}{
 2020-06-17
}

\section{Code Author}{
 generated by the \code{modelbuilder} R package
}

\section{Code creation date}{
 2020-06-17
}

\examples{
 
# To run the simulation with default parameters:  
result <- simulate_SEIRS_model_stochastic() 
# To choose values other than the standard one, specify them like this:  
result <- simulate_SEIRS_model_stochastic( vars = c(S = 2000,E = 2,I = 2,R = 0)) 
# You can display or further process the result, like this:  
plot(result$ts[,'time'],result$ts[,'S'],xlab='Time',ylab='Numbers',type='l') 
print(paste('Max number of S: ',max(result$ts[,'S']))) 
}
