% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_stochastic_SIR.R
\name{simulate_stochastic_SIR}
\alias{simulate_stochastic_SIR}
\title{Stochastic simulation of an SIR-type model}
\usage{
simulate_stochastic_SIR(S0 = 1000, I0 = 10, tmax = 100, b = 1/1000,
  g = 0.5, m = 0, n = 0, rngseed = 100)
}
\arguments{
\item{S0}{initial number of susceptible hosts}

\item{I0}{initial number of infected, symptomatic hosts}

\item{tmax}{maximum simulation time, units depend on choice of units for 
your parameters}

\item{b}{level/rate of infectiousness for hosts in the I compartment}

\item{g}{rate at which a person leaves the I compartment}

\item{m}{the rate at which new individuals enter the model (are born)}

\item{n}{the rate of natural death (the inverse is the average lifespan)}

\item{rngseed}{seed for random number generator to allow reproducibility}
}
\value{
The function returns a list. The list has one element, a data frame ts
which contains the time series of the simulated model, 
with one column per compartment/variable. The first column is time.
}
\description{
Simulation of a stochastic SIR type model with the following
  compartments: Susceptibles (S), 
  Infected and Infectious (I), Recovered and Immune (R)
}
\details{
A compartmental SIR model is
  simulated. Initial conditions for the R variable is 0. Units of
  time depend on the time units chosen for model parameters. The simulation
  runs as a stochastic model using the adaptive-tau algorithm as implemented
  by ssa.adaptivetau in the adpativetau package. See the manual of this
  package for more details. The function returns a list,
  with the time series of the
  simulated disease as list element ts, with one column per
  compartment/variable. The first column is time.
}
\section{Warning}{
 This function does not perform any error checking. So if
  you try to do something nonsensical (e.g. specify negative parameter values
  or fractions > 1), the code will likely abort with an error message
}

\examples{
# To run the simulation with default parameters 
result <- simulate_stochastic_SIR()
# To choose parameter values other than the standard one, specify them e.g. like this
result <- simulate_stochastic_SIR(S0 = 2000,  tmax = 200, b = 1/100)
# You can display or further process the result, e.g. like this
plot(result$ts[,'Time'],result$ts[,'S'],xlab='Time',ylab='Number Susceptible',type='l')
print(paste('Max number of infected: ',max(result$ts[,'I']))) 
}
\seealso{
See the shiny app documentation corresponding to this simulator
function for more details on this model. See the manual for the adaptivetau
package for details on the stochastic algorithm.
}
\author{
Andreas Handel
}
