% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSAIDE.R
\docType{package}
\name{DSAIDE}
\alias{DSAIDE}
\title{DSAIDE: A package to learn about Dynamical Systems Approaches to Infectious
Disease Epidemiology}
\description{
This package provides a number of shiny apps that simulate various
infectious disease dynamics models By manipulating the models and working
through the instructions provided within the shiny UI, you can learn about
some important concepts in infectious disease epidemiology. You will also
learn how models can be used to study such concepts.
}
\details{
Start the main menu of the package by calling dsaidemenu().

To learn more about how to use the package, see the documentation
on the package website:
https://ahgroup.github.io/DSAIDE/
}
