% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_vectortransmission.R
\name{simulate_vectortransmission}
\alias{simulate_vectortransmission}
\title{Simulation of a compartmental infectious disease transmission model illustrating vector-borne transmission}
\usage{
simulate_vectortransmission(S0 = 1000, I0 = 1, Sv0 = 0, tmax = 120,
  beta1 = 0.01, beta2 = 0, beta3 = 0, beta4 = 0, lambda1 = 0,
  lambda2 = 0, n1 = 0, n2 = 0, gamma1 = 1, gamma2 = 1, p = 0,
  c = 0)
}
\arguments{
\item{S0}{initial number of susceptible hosts}

\item{I0}{initial number of infected hosts}

\item{Sv0}{initial number of susceptible vectors}

\item{tmax}{maximum simulation time, units of months}

\item{beta1}{rate of transmission from infected to susceptible host}

\item{beta2}{rate of transmission from infected vector to susceptible host}

\item{beta3}{rate of transmission from environment to susceptible host}

\item{beta4}{rate of transmission from infected host to susceptible vector}

\item{lambda1}{the rate of births of hosts}

\item{lambda2}{the rate of births of vectors}

\item{n1}{the rate of natural death of hosts}

\item{n2}{the rate of natural death of vectors}

\item{gamma1}{the rate at which infected hosts recover/die}

\item{gamma2}{the rate at which infected vectors recover/die}

\item{p}{the rate at which infected host shed pathogen in the enviroment}

\item{c}{the rate at which pathogen in the enviroment decays}
}
\value{
This function returns the simulation result as obtained from a call
  to the deSolve ode solver
}
\description{
This model allows for the simulation of a vector-borne infectious disease
}
\details{
A compartmental ID model with several states/compartments
  is simulated as a set of ordinary differential
  equations. The function returns the output from the odesolver as a matrix,
  with one column per compartment/variable. The first column is time.
}
\section{Warning}{

  This function does not perform any error checking. So if you try to do
  something nonsensical (e.g. any negative values or fractions > 1),
  the code will likely abort with an error message
}
\examples{
  # To run the simulation with default parameters just call this function
  result <- simulate_vectortransmission()
  # To choose parameter values other than the standard one, specify them e.g. like such
  result <- simulate_vectortransmission(S0 = 100, Sv0 = 1e5,  tmax = 100)
  # You should then use the simulation result returned from the function, e.g. like this:
  plot(result[,1],result[,2],xlab='Time',ylab='Number Susceptible',type='l')
}
\author{
Andreas Handel
}
\references{
See e.g. Keeling and Rohani 2008 for SIR models and the
  documentation for the deSolve package for details on ODE solvers
}
\seealso{
The UI of the shiny app 'VectorTransmission', which is part of this package, contains more details on the model
}

